refactor: reorganize project to Go standard structure (internal/executor, internal/runner)
