+++ internal/executor/executor_test.go
@@ -1,4 +1,4 @@
-package main
+package executor
 
 import (
 	"bytes"
