+++ internal/runner/real_runner_test.go
@@ -1,4 +1,5 @@
 package main
+package main
 
 import (
 	"bytes"
